fix: устранить находки code review
Продакшн-код:
- median: переполнение Int при сложении двух больших элементов
- stringSet-делегат: toSet() неизменяем только для 0-1 элемента,
  теперь unmodifiableSet поверх копии
- BracketsValidatorJava: private static final вместо package-private
  изменяемых статиков, приватный конструктор

Тесты:
- кейсы на не-скобочные символы (буква, пробел) - раньше правило
  вообще не проверялось
- ObjectSingletonKotlin: лямбда вместо bound reference, который
  создавал синглтон до входа в проверку
- гонка засчитывается только по конкретному признаку, NPE от
  lazy(NONE) больше не роняет тест
- таймауты в ConcurrencyHarness вместо вечного ожидания
- пул потоков один на класс, ITERATIONS 500 -> 100
- убран хрупкий ассерт на границе ceil (p = 0.9)

Сборка и структура:
- android-стаб без транзитивных зависимостей
- сериализация переведена в testImplementation, версия плагина
  выровнена с Kotlin 2.4.0
- UTF-8 в отчётах тестов
- пакеты brackets_validator/strings_combiner -> brackets/strings
- ListMapperTest -> TransformersTest, StringsCombinerTest -> StringsCombinersTest
- internal для тестовых утилит, переводы строк, битая ссылка в KDoc

+++ src/main/kotlin/org/example/extensions/ListExtensions.kt
@@ -7,6 +7,7 @@ fun List<Int>.median(): Double? {
     return if (size % 2 == 1) {
         sorted[middle].toDouble()
     } else {
-        (sorted[middle - 1] + sorted[middle]) / 2.0
+        // toLong() обязателен: сумма двух больших Int переполнилась бы до расширения до Double
+        (sorted[middle - 1].toLong() + sorted[middle]) / 2.0
     }
 }
